+++ etail_marketplaces_sdk/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility helpers for the etail-marketplaces-sdk."""
+
+from etail_marketplaces_sdk.utils.dataframe import to_dataframe
+
+__all__ = ["to_dataframe"]